oracle/requirements.txt
# Oracle Backend Requirements
# Cloud-native security analytics platform

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
msal==1.26.0
cryptography==41.0.7
PyJWT==2.8.0
google-auth==2.25.2

# HTTP Client for Sentry communication
httpx==0.25.2
requests==2.31.0

# Data Processing & Analytics 
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.3.2

# Azure OpenAI & AI Services
openai==1.12.0
azure-identity==1.15.0

# Monitoring & Observability
prometheus-client==0.19.0
structlog==23.2.0

# Configuration Management
python-decouple==3.8
pyyaml==6.0.1

# Date/Time handling
python-dateutil==2.8.2

# Testing (Fixed typo/non-existent package)
pytest==7.4.3
pytest-asyncio==0.21.1

# Environment Configuration
python-dotenv==1.0.0